Gir opp, god natt

+++ src/Parser.cpp
@@ -394,10 +394,11 @@ auto Parser::onlyVertex() -> Vertex
 
 
     // uv
-    const GLushort u = 65535U * __OPTMIZED__StringToFloat(it, &end);
+    const GLushort u = static_cast<GLushort>( 65535U * __OPTMIZED__StringToFloat(it, &end));
     it = end;
-    const GLushort v = 65535U * __OPTMIZED__StringToFloat(it, &end);
+    const GLushort v = static_cast<GLushort>( 65535U * __OPTMIZED__StringToFloat(it, &end));
     it = end;
+
 
 
     // @optmization Function will only work here. Not genral at all
